Remove a simple warning

+++ Source/MediaStorageAndFileFormat/gdcmDictPrinter.cxx
@@ -80,8 +80,8 @@ VM GuessVMType(DataElement const &de)
           {
           assert( bv && "not bv" );
           const char *array = bv->GetPointer();
-          unsigned int c = VM::GetNumberOfElementsFromArray(array, vl);
-          vm = VM::GetVMTypeFromLength( c, 1 );
+          size_t c = VM::GetNumberOfElementsFromArray(array, vl);
+          vm = VM::GetVMTypeFromLength( (unsigned int)c, 1 );
           }
         }
       break;
